Updated TextCaseConverter
