+++ XamarinCommunityToolkit/Converters/TextCaseConverterType.shared.cs
@@ -1,0 +1,13 @@
+﻿
+namespace XamarinCommunityToolkit.Converters
+{
+    /// <summary>
+    /// The text case that the TextCaseConverter should convert a value.
+    /// </summary>
+    public enum TextCaseConverterType
+    {
+        None,
+        Upper,
+        Lower,
+    }
+}
